Feature/LAP203-23 us101   funcao gere nrs aleatorios

+++ ARQCP/US101/demo_pcg.h
@@ -1,6 +1,6 @@
 #ifndef DEMO_PCG_H
 #define DEMO_PCG_H
 
-	unsigned int pcg32_random_r();
+	unsigned long long pcg32_random_r();
 
 #endif
